requirements.txt
# FastAPI和Web服务依赖 - Python 3.12兼容版本
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
starlette>=0.27.0

# 数据验证和环境变量
pydantic>=2.4.0,<3.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# LangChain相关依赖 - 最新稳定版本
langchain>=0.1.0,<0.2.0
langchain-core>=0.1.0,<0.2.0
langchain-openai>=0.1.0,<0.2.0
langchain-community>=0.0.20,<0.1.0
langchain-text-splitters>=0.0.1

# OpenAI API
openai>=1.30.0
tiktoken>=0.5.0

# 向量数据库和科学计算
faiss-cpu>=1.8.0
chromadb>=0.4.0
numpy>=1.24.0,<2.0.0
scipy>=1.10.0

# 文档处理
pypdf2>=3.0.0
markdown>=3.4.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# 其他依赖
requests>=2.31.0
aiofiles>=23.0.0
typing-extensions>=4.8.0